+++ js/example/get-random-positive-integer.js
